requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
qdrant-client==1.6.9
sentence-transformers==2.2.2
rank-bm25==0.2.2
jieba==0.42.1
PyPDF2==3.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
lxml==4.9.3
